apps/posts/serializers.py
from posts.models import Post
from rest_framework import serializers


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ('id', 'user', 'text')
        read_only_fields = ('id',)
